refactor: move MCPConfigDialog to its own  MCP subdir in ui and update import
