+++ client/src/components/ui/MCP/MCPConfigDialog.tsx
